Refactor graph methods to separate module.
